Final solution
I probably should have committed along the way! Code evolved a lot before it
got to here. There's a functional / procedural approach here, the examples
growing in complexity meant it was easy to encapsulate behaviours and extract
them.

Initially there were a lot of arguments to functions before I encapsulated them
in an object, it made it much easier to read what was going on when being given
to functions.

As things went on I introduced more of a functional style, with currying the
working days behaviour and advancing days.

Was fun!